Se pasa a componente la carta de producto
